perf(bam): shrink transaction storage and index tracked batches
Keep fd_txn_p_t at 4992 bytes with ingress timestamps in Pack sidecar
state. Index live BAM signatures and cap outstanding work at 8192
batches, with twice that capacity for results.

Enable BAM certificate verification and reduce BAM/bundle TLS heap
defaults to 8 MiB. Retry BAM BIO/SSL creation failures and count failed
OpenSSL reallocations. Tighten comments and drop redundant log arguments
and the single-use insertion helper.

Validation: firedancer-dev and fdctl builds; 13 focused unit tests.
Pack BAM tests rerun after the final helper cleanup.

+++ src/flamenco/runtime/tests/fd_instr_harness.c
@@ -67,7 +67,7 @@ fd_solfuzz_pb_instr_ctx_create( fd_solfuzz_runner_t *                runner,
   fd_txn_t *   txn_descriptor = TXN( txn );
   if( test_ctx->data ) {
     memcpy( txn->payload, test_ctx->data->bytes, test_ctx->data->size );
-    txn->payload_sz = test_ctx->data->size;
+    txn->payload_sz = (ushort)(test_ctx->data->size);
   } else {
     txn->payload_sz = 0;
   }
